Consolidate eligibility logic
